Update Compression unit tests.
